MINOR: Move a few more methods to AuthHelper
Also move some tests to `AuthHelperTest` and fix
logger used by `CoreUtils.swallow`.
