+++ core/src/main/scala/kafka/server/AuthHelper.scala
@@ -19,9 +19,9 @@ package kafka.server
 
 import java.lang.{Byte => JByte}
 import java.util.Collections
-
 import kafka.network.RequestChannel
 import kafka.security.authorizer.AclEntry
+import kafka.utils.CoreUtils
 import org.apache.kafka.common.acl.AclOperation
 import org.apache.kafka.common.errors.ClusterAuthorizationException
 import org.apache.kafka.common.requests.RequestContext
@@ -31,8 +31,8 @@ import org.apache.kafka.common.resource.{PatternType, Resource, ResourcePattern,
 import org.apache.kafka.common.utils.Utils
 import org.apache.kafka.server.authorizer.{Action, AuthorizationResult, Authorizer}
 
+import scala.collection.{Map, Seq}
 import scala.jdk.CollectionConverters._
-
 
 class AuthHelper(authorizer: Option[Authorizer]) {
 
@@ -70,4 +70,64 @@ class AuthHelper(authorizer: Option[Authorizer]) {
     }
     Utils.to32BitField(authorizedOps.map(operation => operation.code.asInstanceOf[JByte]).asJava)
   }
+
+  def authorizeByResourceType(requestContext: RequestContext, operation: AclOperation,
+                              resourceType: ResourceType): Boolean = {
+    authorizer.forall { authZ =>
+      authZ.authorizeByResourceType(requestContext, operation, resourceType) == AuthorizationResult.ALLOWED
+    }
+  }
+
+  def partitionSeqByAuthorized[T](requestContext: RequestContext,
+                                  operation: AclOperation,
+                                  resourceType: ResourceType,
+                                  resources: Seq[T],
+                                  logIfAllowed: Boolean = true,
+                                  logIfDenied: Boolean = true)(resourceName: T => String): (Seq[T], Seq[T]) = {
+    authorizer match {
+      case Some(_) =>
+        val authorizedResourceNames = filterByAuthorized(requestContext, operation, resourceType,
+          resources, logIfAllowed, logIfDenied)(resourceName)
+        resources.partition(resource => authorizedResourceNames.contains(resourceName(resource)))
+      case None => (resources, Seq.empty)
+    }
+  }
+
+  def partitionMapByAuthorized[K, V](requestContext: RequestContext,
+                                     operation: AclOperation,
+                                     resourceType: ResourceType,
+                                     resources: Map[K, V],
+                                     logIfAllowed: Boolean = true,
+                                     logIfDenied: Boolean = true)(resourceName: K => String): (Map[K, V], Map[K, V]) = {
+    authorizer match {
+      case Some(_) =>
+        val authorizedResourceNames = filterByAuthorized(requestContext, operation, resourceType,
+          resources.keySet, logIfAllowed, logIfDenied)(resourceName)
+        resources.partition { case (k, _) => authorizedResourceNames.contains(resourceName(k)) }
+      case None => (resources, Map.empty)
+    }
+  }
+
+  def filterByAuthorized[T](requestContext: RequestContext,
+                                            operation: AclOperation,
+                                            resourceType: ResourceType,
+                                            resources: Iterable[T],
+                                            logIfAllowed: Boolean = true,
+                                            logIfDenied: Boolean = true)(resourceName: T => String): Set[String] = {
+    authorizer match {
+      case Some(authZ) =>
+        val resourceNameToCount = CoreUtils.groupMapReduce(resources)(resourceName)(_ => 1)(_ + _)
+        val actions = resourceNameToCount.iterator.map { case (resourceName, count) =>
+          val resource = new ResourcePattern(resourceType, resourceName, PatternType.LITERAL)
+          new Action(operation, resource, count, logIfAllowed, logIfDenied)
+        }.toBuffer
+        authZ.authorize(requestContext, actions.asJava).asScala
+          .zip(resourceNameToCount.keySet)
+          .collect { case (authzResult, resourceName) if authzResult == AuthorizationResult.ALLOWED =>
+            resourceName
+          }.toSet
+      case None => resources.iterator.map(resourceName).toSet
+    }
+  }
+
 }
